+++ src/headers/ProfResults.h
@@ -22,12 +22,12 @@ namespace mpplas{
       ProfResults& operator+=(const ProfResults& rhs);
       ProfResults& operator-=(const ProfResults& rhs);
 
-      const ProfResult& operator[](const size_t thread) const;
-      ProfResult& operator[](const size_t thread) ;
+      const ProfResult& operator[](const int thread) const;
+      ProfResult& operator[](const int thread) ;
 
       OpsCount getTotalOps() const;
 
-      size_t size() const;
+      int size() const;
 
 
     private:
